Add tooltip and hover/focus state
